Further desctructering and JSDOC, need to look further into listing-detail-render.mjs and divide because of the volume of code.
